requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
sqlalchemy
psycopg2-binary
passlib
bcrypt==3.2.0
python-jose[cryptography]
python-multipart
google-api-python-client
google-auth-oauthlib
pyasn1==0.4.8